pyproject.toml
[tool.poetry]
name = "coop-tickets"
version = "0.1.0"
description = "Small Windows terminal Python application to process Coop supermarket pdf tickets into Excel files"
authors = ["Eufemiano Fuentes <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.0"
pandas = "^2.1.4"
tabula-py = {extras = ["jpype"], version = "^2.9.0"}
openpyxl = "^3.1.2"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
